binarySearch class created.

+++ DCMS/src/main/java/com/nima/dcms/urlconverter/MD5URLConverter.kt
@@ -5,7 +5,7 @@ import java.security.MessageDigest
 
 
 class MD5URLConverter : URLConverter {
-    override fun convert(url: String): Int {
+    override fun convert(url: String): Long {
         val md5: MessageDigest = MessageDigest.getInstance("MD5")
         md5.update(StandardCharsets.UTF_8.encode(url))
 //        return java.lang.String.format("%032x", BigInteger(1, md5.digest()))
